+++ src/main/java/kitchenpos/order/infrastructure/JdbcTemplateOrderDao.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.infrastructure;
 
 import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.dao.OrderDao;
+import kitchenpos.order.domain.repository.OrderDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
